backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
face-recognition==1.3.0
opencv-python==4.5.4.60
numpy==1.21.4
pillow==8.4.0
#bson==0.5.10
PyJWT==2.3.0 